+++ app/__init__.py
@@ -14,6 +14,4 @@ Session(app)
 # Initialize login
 # login = LoginManager(app)
 
-#from app import helpers
 from app import routes
-#from helpers import login_required
