--- app/helpers.py
@@ -1,12 +1,0 @@
-from functools import wraps
-from flask import redirect, render_template, request, session
-
-# this is the logic behind the login_required decorator
-# https://flask.palletsprojects.com/en/1.1.x/patterns/viewdecorators/
-def login_required(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-        if session.get("user_id") is None:
-            return redirect(url_for('login', next=request.url))
-        return f(*args, **kwargs)
-    return decorated_function
